requirements_test.txt
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
Django==2.1.15
django-crispy-forms==1.7.2
django-phonenumber-field==2.3.1
django-twilio==0.10.0
gunicorn==19.9.0
idna==2.8
phonenumbers==8.10.8
PyJWT==1.7.1
PySocks==1.6.8
pytz==2018.9
requests==2.21.0
six==1.12.0
twilio==6.26.0
urllib3==1.24.2
coverage==5.0.2